install apollo, fix header
