+++ package.json
@@ -3,13 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "apollo-cache-inmemory": "^1.1.11",
+    "apollo-client": "^2.2.7",
+    "apollo-link": "^1.2.1",
+    "apollo-link-http": "^1.5.3",
+    "apollo-link-state": "^0.4.1",
     "classnames": "^2.2.5",
     "formik": "^0.11.11",
+    "graphql": "^0.13.1",
+    "graphql-tag": "^2.8.0",
     "immutable": "^3.8.2",
     "js-cookie": "^2.2.0",
     "prettier": "^1.11.1",
     "query-string": "^5.1.0",
     "react": "^16.2.0",
+    "react-apollo": "^2.1.0-beta.3",
     "react-document-title": "^2.0.3",
     "react-dom": "^16.2.0",
     "react-router-dom": "^4.2.2",
